% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hsrs.R
\name{create_hsrs}
\alias{create_hsrs}
\title{Create the matrix of half-saturation constants
  of species (rows) on resources (columns)}
\usage{
create_hsrs(n_species, n_resources)
}
\arguments{
\item{n_species}{number of species}

\item{n_resources}{number of resources}
}
\description{
Create the matrix of half-saturation constants
  of species (rows) on resources (columns)
}
\author{
Richel J.C. Bilderbeek
}
