% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hsrs.R
\name{create_hsrs_1_a}
\alias{create_hsrs_1_a}
\title{Create the matrix of half-saturation constants
  of 3 species (columns) on 3 resources (rows)
  as used in figure 1 A}
\usage{
create_hsrs_1_a()
}
\description{
Create the matrix of half-saturation constants
  of 3 species (columns) on 3 resources (rows)
  as used in figure 1 A
}
\author{
Richel J.C. Bilderbeek
}
