% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hsrs.R
\name{create_hsrs_1_c}
\alias{create_hsrs_1_c}
\title{Create the matrix of half-saturation constants
  of 6 species (columns) on 3 resources (rows)
  as used in igure 1C}
\usage{
create_hsrs_1_c()
}
\description{
Create the matrix of half-saturation constants
  of 6 species (columns) on 3 resources (rows)
  as used in igure 1C
}
\author{
Richel J.C. Bilderbeek
}
