% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hsrs.R
\name{create_hsrs_1_d}
\alias{create_hsrs_1_d}
\title{Create the matrix of half-saturation constants
  of 6 species (columns) on 3 resources (rows)
  as used in figure 1 D}
\usage{
create_hsrs_1_d()
}
\description{
Create the matrix of half-saturation constants
  of 6 species (columns) on 3 resources (rows)
  as used in figure 1 D
}
\author{
Richel J.C. Bilderbeek
}
