% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_spec_densities.R
\name{create_init_spec_densities}
\alias{create_init_spec_densities}
\title{Create the initial species densities for the species present at time zero}
\usage{
create_init_spec_densities(n_species)
}
\description{
Create the initial species densities for the species present at time zero
}
\author{
Richel J.C. Bilderbeek
}
