% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_srcs.R
\name{create_srcs}
\alias{create_srcs}
\title{Create the matrix of species' (rows) resource (columns) content}
\usage{
create_srcs(n_species, n_resources)
}
\arguments{
\item{n_species}{number of species}

\item{n_resources}{number of resources}
}
\description{
Create the matrix of species' (rows) resource (columns) content
}
\author{
Richel J.C. Bilderbeek
}
