% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_srcs.R
\name{create_srcs_1_a}
\alias{create_srcs_1_a}
\title{Create the species resource contents used in figure 1 A.}
\usage{
create_srcs_1_a()
}
\description{
Create the species resource contents used in figure 1 A.
}
\author{
Richel J.C. Bilderbeek
}
