% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_srcs.R
\name{create_srcs_1_c}
\alias{create_srcs_1_c}
\title{Create the species resource contents used in figure 1 C.}
\usage{
create_srcs_1_c()
}
\description{
Create the species resource contents used in figure 1 C.
}
\author{
Richel J.C. Bilderbeek
}
