% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_srcs.R
\name{create_srcs_1_d}
\alias{create_srcs_1_d}
\title{Create the species resource contents used in figure 1 D.}
\usage{
create_srcs_1_d()
}
\description{
Create the species resource contents used in figure 1 D.
}
\author{
Richel J.C. Bilderbeek
}
