% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_uptake_matrix.R
\name{create_uptake_matrix}
\alias{create_uptake_matrix}
\title{Create the uptake matrix, the right-hand part of equation 2}
\usage{
create_uptake_matrix(srcs, species_densities, resource_densities,
  max_growth_rates, hsrs, mortality_rates)
}
\arguments{
\item{species_densities}{density of all species}

\item{resource_densities}{resource densities}

\item{max_growth_rates}{max growth rate of all species}

\item{hsrs}{a matrix of half-saturation constants
of species (rows) on resources (columns). Can be created by
\link{create_hsrs}}

\item{mortality_rates}{mortality rates of all species}
}
\description{
Create the uptake matrix, the right-hand part of equation 2
}
\author{
Richel J.C. Bilderbeek
}
