% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(hsr, hsrs, max_growth_rate, n_species, n_resources,
  resource_densities)
}
\arguments{
\item{hsr}{half-saturation constants for
a focal species for each of the resources}

\item{hsrs}{a matrix of half-saturation constants
of species (rows) on resources (columns). Can be created by
\link{create_hsrs}}

\item{max_growth_rate}{max growth rate of the focal species}

\item{n_species}{number of species}

\item{n_resources}{number of resources}

\item{resource_densities}{resource densities}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
