% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_growth_rates.R
\name{get_max_growth_rates}
\alias{get_max_growth_rates}
\title{Get the maximum growth rates of a number of species}
\usage{
get_max_growth_rates(n_species)
}
\description{
Get the maximum growth rates of a number of species
}
\author{
Richel J.C. Bilderbeek
}
