% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mortality_rates.R
\name{get_mortality_rates}
\alias{get_mortality_rates}
\title{Get the mortality rates for any number of species}
\usage{
get_mortality_rates(n_species)
}
\arguments{
\item{n_species}{number of species}
}
\description{
Get the mortality rates for any number of species
}
\author{
Richel J.C. Bilderbeek
}
