% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run the simulation}
\usage{
run_sim(initial_species_densities, initial_resource_densities, dt, t_max)
}
\value{
a data frame
  with the species densities and resource concentrations (as columns)
  in time (as rows)
}
\description{
Run the simulation
}
\author{
Richel J.C. Bilderbeek
}
