% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resource_changes.R
\name{calc_resource_changes}
\alias{calc_resource_changes}
\title{Calculate the change in resource, equation 2}
\usage{
calc_resource_changes(max_growth_rates, resource_densities, hsrs, srcs,
  species_densities, supply_rates, turnover_rate, mortality_rates)
}
\description{
Calculate the change in resource, equation 2
}
\author{
Richel J.C. Bilderbeek
}
