% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_specific_growth_rate.R
\name{calc_specific_growth_rate}
\alias{calc_specific_growth_rate}
\title{Calculate the specific growth rates, equation 3}
\usage{
calc_specific_growth_rate(resource_densities, max_growth_rate,
  half_saturation_resources)
}
\description{
Calculate the specific growth rates, equation 3
}
\author{
Richel J.C. Bilderbeek
}
