% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(dt, hsr, hsrs, initial_resource_densities,
  initial_species_densities, max_growth_rate, max_growth_rates, mortality_rates,
  n_species, n_resources, resource_densities, species_densities, t_max)
}
\arguments{
\item{dt}{delta timestep}

\item{hsr}{half-saturation constants for
a focal species for each of the resources}

\item{hsrs}{a matrix of half-saturation constants
of species (rows) on resources (columns). Can be created by
\link{create_hsrs}}

\item{initial_resource_densities}{the resource densities at time zero}

\item{initial_species_densities}{the species densities at time zero}

\item{max_growth_rate}{max growth rate of the focal species}

\item{max_growth_rates}{max growth rate of all species}

\item{mortality_rates}{mortality rates of all species}

\item{n_species}{number of species}

\item{n_resources}{number of resources}

\item{resource_densities}{resource densities}

\item{species_densities}{density of all species}

\item{t_max}{simulation will run in steps of \code{dt}
until (not inclusive) this time point is reached}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richel J.C. Bilderbeek
}
